Move signer capability into Nostr signing
