+++ web/src/lib/nostr/signing/signer-capability.ts
