+++ web/src/lib/signer-strategy.test.ts
@@ -1,6 +1,5 @@
 import { describe, it, expect, vi, afterEach } from 'vitest';
 import { resolveSigner } from './signer-strategy';
-import { signerCanSign } from './signer-capability';
 
 function stubLogin(value: string | null): void {
 	vi.stubGlobal('localStorage', {
@@ -13,19 +12,6 @@ function stubLogin(value: string | null): void {
 
 afterEach(() => {
 	vi.unstubAllGlobals();
-});
-
-describe('signerCanSign', () => {
-	it('returns true for signing capable types', () => {
-		expect(signerCanSign('NIP-07')).toBe(true);
-		expect(signerCanSign('NIP-46')).toBe(true);
-		expect(signerCanSign('nsec')).toBe(true);
-	});
-
-	it('returns false for npub and undefined', () => {
-		expect(signerCanSign('npub')).toBe(false);
-		expect(signerCanSign(undefined)).toBe(false);
-	});
 });
 
 describe('resolveSigner', () => {
